requirements.txt
flask==3.0.0
pdfplumber==0.10.3
python-docx==1.1.0
google-generativeai>=0.8.0
fpdf==1.7.2
werkzeug==3.0.1
